fix: Update Cart

+++ backend/src/main/java/toughcircle/shop/service/CartService.java
@@ -2,12 +2,12 @@ package toughcircle.shop.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.coyote.BadRequestException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import toughcircle.shop.model.Entity.Cart;
 import toughcircle.shop.model.Entity.CartItem;
 import toughcircle.shop.model.Entity.Product;
+import toughcircle.shop.model.Entity.User;
 import toughcircle.shop.model.dto.CartItemDto;
 import toughcircle.shop.model.dto.request.AddCartItemRequest;
 import toughcircle.shop.model.dto.request.UpdateQuantityRequest;
@@ -16,6 +16,7 @@ import toughcircle.shop.repository.CartRepository;
 import toughcircle.shop.repository.ProductRepository;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -29,15 +30,18 @@ public class CartService {
     private final ProductService productService;
     private final TokenUserService tokenUserService;
 
+    /**
+     * 장바구니에 상품 추가
+     * @param cartId 장바구니 ID
+     * @param request 장바구니 상품 추가 요청 정보
+     */
     @Transactional
-    public void saveCartItem(String token, Long cartId, AddCartItemRequest request) throws BadRequestException {
-        tokenUserService.getUserByToken(token);
-
+    public void saveCartItem(Long cartId, AddCartItemRequest request) {
         Cart cart = cartRepository.findById(cartId)
-            .orElseThrow(() -> new BadRequestException("Cart not found with cartId: " + cartId));
+            .orElseThrow(() -> new RuntimeException("Cart not found with cartId: " + cartId));
 
         Product product = productRepository.findById(request.getProductId())
-            .orElseThrow(() -> new BadRequestException("Product not found with productId: " + request.getProductId()));
+            .orElseThrow(() -> new RuntimeException("Product not found with productId: " + request.getProductId()));
 
         CartItem item = cartItemRepository.findByCart_idAndProduct_id(cartId, request.getProductId());
         if (item != null) {
@@ -54,32 +58,70 @@ public class CartService {
 
     }
 
+    /**
+     * 장바구니 아이템의 수량 업데이트.
+     * @param cartItemId 장바구니 아이템 ID
+     * @param request 수량 업데이트 요청 정보
+     */
     @Transactional
-    public void updateQuantity(String token, Long cartItemId, UpdateQuantityRequest request) throws BadRequestException {
-        tokenUserService.getUserByToken(token);
+    public void updateQuantity(String token, Long cartItemId, UpdateQuantityRequest request) {
+        User user = tokenUserService.getUserByToken(token);
 
         CartItem cartItem = cartItemRepository.findById(cartItemId)
-            .orElseThrow(() -> new BadRequestException("CartItem not found with cartItemId: " + cartItemId));
+            .orElseThrow(() -> new RuntimeException("CartItem not found with cartItemId: " + cartItemId));
 
-        cartItem.setQuantity(request.getQuantity());
+        if (Objects.equals(cartItem.getCart().getId(), user.getCart().getId())) {
+            cartItem.setQuantity(request.getQuantity());
+            cartItemRepository.save(cartItem);
+        } else {
+            throw new RuntimeException("User is not authorized to update this cart item.");
+        }
     }
 
-    public void deleteCartItem(String token, Long cartItemId) throws BadRequestException {
-        tokenUserService.getUserByToken(token);
+    /**
+     * 장바구니 아이템 삭제
+     * @param token 사용자 토큰
+     * @param cartItemId 장바구니 아이템 ID
+     */
+    public void deleteCartItem(String token, Long cartItemId) {
+        User user = tokenUserService.getUserByToken(token);
 
-        cartItemRepository.deleteById(cartItemId);
+        CartItem cartItem = cartItemRepository.findById(cartItemId)
+            .orElseThrow(() -> new RuntimeException("CartItem not found with cartItemId: " + cartItemId));
+
+        if (Objects.equals(cartItem.getCart().getId(), user.getCart().getId())) {
+            cartItemRepository.deleteById(cartItemId);
+        } else {
+            throw new RuntimeException("User is not authorized to delete this cart item.");
+        }
+
     }
 
-    public List<CartItemDto> getCartItem(String token, Long cartId) throws BadRequestException {
+    /**
+     * 장바구니의 상품 목록 조회
+     * @param token 사용자 토큰
+     * @param cartId 장바구니 ID
+     * @return 장바구니 아이템 DTO 리스트
+     */
+    public List<CartItemDto> getCartItems(String token, Long cartId) {
+        User user = tokenUserService.getUserByToken(token);
+
         Cart cart = cartRepository.findById(cartId)
-            .orElseThrow(() -> new BadRequestException("Cart not found with cartId: " + cartId));
+            .orElseThrow(() -> new RuntimeException("Cart not found with cartId: " + cartId));
 
-        List<CartItem> cartItemList = cart.getCartItemList();
-        List<CartItemDto> list = cartItemList.stream().map(this::convertToDto).toList();
-        return list;
-
+        if (Objects.equals(cart.getId(), user.getCart().getId())) {
+            List<CartItem> cartItems = cart.getCartItemList();
+            return cartItems.stream().map(this::convertToDto).toList();
+        } else {
+            throw new RuntimeException("User is not authorized to view this cart.");
+        }
     }
 
+    /**
+     * 장바구니 아이템 DTO 변환
+     * @param cartItem 장바구니 아이템 엔티티
+     * @return 장바구니 아이템 DTO
+     */
     private CartItemDto convertToDto(CartItem cartItem) {
         CartItemDto cartItemDto = new CartItemDto();
         cartItemDto.setCartItemId(cartItem.getId());
